refactor(materializer): extract unresolved-tool-diagnostics (file #3)
Lift the opt-in unresolved-tool diagnostic logging (gated behind the
acepe:debug:unresolved-tools flag) into its own module. Read-only over the
canonical source. Net 74/74, check 0.
